[router-server] Separate out `getNamedParametrizedRoute`
