+++ packages/@expo/router-server/src/__tests__/getNamedParametrizedRoute.test.node.ts
@@ -1,0 +1,11 @@
+import { parseParameter } from '../getNamedParametrizedRoute';
+
+describe(parseParameter, () => {
+  it(`matches optionals using non-standard from router v1`, () => {
+    expect(parseParameter('[...all]')).toEqual({
+      name: 'all',
+      optional: true,
+      repeat: true,
+    });
+  });
+});
